# Introduce WASM JITs for interpreter opcodes, do_jit_call, and interp_entry wrappers