--- a/src/mono/mono/mini/interp/interp.c
+++ b/src/mono/mono/mini/interp/interp.c
@@ -82,6 +82,11 @@
 #endif
 #include <mono/metadata/icall-decl.h>
 
+#ifdef HOST_BROWSER
+#include "jiterpreter.h"
+#include <emscripten.h>
+#endif
+
 /* Arguments that are passed when invoking only a finally/filter clause from the frame */
 struct FrameClauseArgs {
 	/* Where we start the frame execution from */
@@ -247,9 +252,6 @@ static gboolean interp_init_done = FALSE;
 #ifdef HOST_WASI
 static gboolean debugger_enabled = FALSE;
 #endif
-
-static void
-interp_exec_method (InterpFrame *frame, ThreadContext *context, FrameClauseArgs *clause_args);
 
 static MonoException* do_transform_method (InterpMethod *imethod, InterpFrame *method, ThreadContext *context);
 
@@ -1168,7 +1170,7 @@ INTERP_GET_EXCEPTION_CHAR_ARG(argument_out_of_range)
 		}									\
 	} while (0)
 
-// Reduce duplicate code in interp_exec_method
+// Reduce duplicate code in mono_interp_exec_method
 static MONO_NEVER_INLINE void
 do_safepoint (InterpFrame *frame, ThreadContext *context, const guint16 *ip)
 {
@@ -2113,13 +2115,13 @@ interp_runtime_invoke (MonoMethod *method, void *obj, void **params, MonoObject 
 	// The method to execute might not be transformed yet, so we don't know how much stack
 	// it uses. We bump the stack_pointer here so any code triggered by method compilation
 	// will not attempt to use the space that we used to push the args for this method.
-	// The real top of stack for this method will be set in interp_exec_method once the
+	// The real top of stack for this method will be set in mono_interp_exec_method once the
 	// method is transformed.
 	context->stack_pointer = (guchar*)(sp + 4);
 	g_assert (context->stack_pointer < context->stack_end);
 
 	MONO_ENTER_GC_UNSAFE;
-	interp_exec_method (&frame, context, NULL);
+	mono_interp_exec_method (&frame, context, NULL);
 	MONO_EXIT_GC_UNSAFE;
 
 	context->stack_pointer = (guchar*)sp;
@@ -2224,7 +2226,7 @@ interp_entry (InterpEntryData *data)
 	g_assert (context->stack_pointer < context->stack_end);
 
 	MONO_ENTER_GC_UNSAFE;
-	interp_exec_method (&frame, context, NULL);
+	mono_interp_exec_method (&frame, context, NULL);
 	MONO_EXIT_GC_UNSAFE;
 
 	context->stack_pointer = (guchar*)sp;
@@ -2521,6 +2523,10 @@ struct _JitCallInfo {
 	gint32 res_size;
 	int ret_mt;
 	gboolean no_wrapper;
+#if HOST_BROWSER
+	int hit_count;
+	WasmJitCallThunk jiterp_thunk;
+#endif
 };
 
 static MONO_NEVER_INLINE void
@@ -2614,11 +2620,19 @@ init_jit_call_info (InterpMethod *rmethod, MonoError *error)
 	rmethod->jit_call_info = cinfo;
 }
 
+#if HOST_BROWSER
+EMSCRIPTEN_KEEPALIVE void
+mono_jiterp_register_jit_call_thunk (void *cinfo, WasmJitCallThunk thunk) {
+	((JitCallInfo*)cinfo)->jiterp_thunk = thunk;
+}
+#endif
+
 static MONO_NEVER_INLINE void
 do_jit_call (ThreadContext *context, stackval *ret_sp, stackval *sp, InterpFrame *frame, InterpMethod *rmethod, MonoError *error)
 {
 	MonoLMFExt ext;
 	JitCallInfo *cinfo;
+	gboolean thrown = FALSE;
 
 	//printf ("jit_call: %s\n", mono_method_full_name (rmethod->method, 1));
 
@@ -2630,7 +2644,49 @@ do_jit_call (ThreadContext *context, stackval *ret_sp, stackval *sp, InterpFrame
 		init_jit_call_info (rmethod, error);
 		mono_error_assert_ok (error);
 	}
+
 	cinfo = (JitCallInfo*)rmethod->jit_call_info;
+
+#if JITERPRETER_ENABLE_JIT_CALL_TRAMPOLINES
+	// FIXME: thread safety
+	if (jiterpreter_jit_call_enabled) {
+		WasmJitCallThunk thunk = cinfo->jiterp_thunk;
+		if (thunk) {
+			MonoFtnDesc ftndesc = {0};
+			void *extra_arg;
+			ftndesc.addr = cinfo->addr;
+			ftndesc.arg = cinfo->extra_arg;
+			extra_arg = cinfo->no_wrapper ? cinfo->extra_arg : &ftndesc;
+			interp_push_lmf (&ext, frame);
+			if (
+				jiterpreter_wasm_eh_enabled ||
+				(mono_aot_mode != MONO_AOT_MODE_LLVMONLY_INTERP)
+			) {
+				thunk (extra_arg, ret_sp, sp, &thrown);
+			} else {
+				mono_interp_invoke_wasm_jit_call_trampoline (
+					thunk, extra_arg, ret_sp, sp, &thrown
+				);
+			}
+			interp_pop_lmf (&ext);
+			goto epilogue;
+		} else {
+			int count = cinfo->hit_count;
+			if (count == JITERPRETER_JIT_CALL_TRAMPOLINE_HIT_COUNT) {
+				void *fn = cinfo->no_wrapper ? cinfo->addr : cinfo->wrapper;
+				mono_interp_jit_wasm_jit_call_trampoline (
+					rmethod, cinfo, fn, rmethod->hasthis, rmethod->param_count,
+					rmethod->arg_offsets, mono_aot_mode == MONO_AOT_MODE_LLVMONLY_INTERP
+				);
+			} else {
+				if (count <= JITERPRETER_JIT_CALL_QUEUE_FLUSH_THRESHOLD)
+					cinfo->hit_count++;
+				if (count == JITERPRETER_JIT_CALL_QUEUE_FLUSH_THRESHOLD)
+					mono_interp_flush_jitcall_queue ();
+			}
+		}
+	}
+#endif
 
 	/*
 	 * Convert the arguments on the interpreter stack to the format expected by the gsharedvt_out wrapper.
@@ -2669,14 +2725,32 @@ do_jit_call (ThreadContext *context, stackval *ret_sp, stackval *sp, InterpFrame
 	}
 
 	interp_push_lmf (&ext, frame);
-	gboolean thrown = FALSE;
+
 	if (mono_aot_mode == MONO_AOT_MODE_LLVMONLY_INTERP) {
+#if JITERPRETER_ENABLE_SPECIALIZED_JIT_CALL
+		/*
+		 * invoke jit_call_cb via a single indirect function call that dispatches to
+		 *  either a specialized JS implementation or a specialized WASM EH version
+		 * see jiterpreter-jit-call.ts and do-jit-call.wat
+		 * NOTE: the first argument must ALWAYS be jit_call_cb for the specialization.
+		 *  the actual implementation cannot verify this at runtime, so get it right
+		 * this is faster than mono_llvm_cpp_catch_exception by avoiding the use of
+		 *  emscripten invoke_vi to find and invoke jit_call_cb indirectly
+		 */
+		jiterpreter_do_jit_call(jit_call_cb, &cb_data, &thrown);
+#else
 		/* Catch the exception thrown by the native code using a try-catch */
 		mono_llvm_cpp_catch_exception (jit_call_cb, &cb_data, &thrown);
+#endif
 	} else {
 		jit_call_cb (&cb_data);
 	}
+
 	interp_pop_lmf (&ext);
+
+#if JITERPRETER_ENABLE_JIT_CALL_TRAMPOLINES
+epilogue:
+#endif
 	if (thrown) {
 		if (context->has_resume_state)
 			/*
@@ -3002,7 +3076,7 @@ interp_entry_from_trampoline (gpointer ccontext_untyped, gpointer rmethod_untype
 	g_assert (context->stack_pointer < context->stack_end);
 
 	MONO_ENTER_GC_UNSAFE;
-	interp_exec_method (&frame, context, NULL);
+	mono_interp_exec_method (&frame, context, NULL);
 	MONO_EXIT_GC_UNSAFE;
 
 	context->stack_pointer = (guchar*)sp;
@@ -3094,7 +3168,7 @@ no_llvmonly_interp_method_pointer (void)
 static MonoFtnDesc*
 interp_create_method_pointer_llvmonly (MonoMethod *method, gboolean unbox, MonoError *error)
 {
-	gpointer addr, entry_func, entry_wrapper;
+	gpointer addr, entry_func = NULL, entry_wrapper;
 	MonoMethodSignature *sig;
 	MonoMethod *wrapper;
 	InterpMethod *imethod;
@@ -3145,10 +3219,33 @@ interp_create_method_pointer_llvmonly (MonoMethod *method, gboolean unbox, MonoE
 	}
 	g_assert (entry_func);
 
+#if HOST_BROWSER
+	// FIXME: We don't support generating wasm trampolines for high arg counts yet
+	if (
+		(sig->param_count <= MAX_INTERP_ENTRY_ARGS) &&
+		jiterpreter_interp_entry_enabled
+	) {
+		jiterp_preserve_module();
+
+		const char *name = mono_method_full_name (method, FALSE);
+		gpointer wasm_entry_func = mono_interp_jit_wasm_entry_trampoline (
+			imethod, method, sig->param_count, (MonoType *)sig->params,
+			unbox, sig->hasthis, sig->ret->type != MONO_TYPE_VOID,
+			name, entry_func
+		);
+		g_free((void *)name);
+
+		// Compiling a trampoline can fail for various reasons, so in that case we will fall back to the pre-existing ones below
+		if (wasm_entry_func)
+			entry_func = wasm_entry_func;
+	}
+#endif
+
 	/* Encode unbox in the lower bit of imethod */
 	gpointer entry_arg = imethod;
 	if (unbox)
 		entry_arg = (gpointer)(((gsize)entry_arg) | 1);
+
 	MonoFtnDesc *entry_ftndesc = mini_llvmonly_create_ftndesc (method, entry_func, entry_arg);
 
 	addr = mini_llvmonly_create_ftndesc (method, entry_wrapper, entry_ftndesc);
@@ -3600,8 +3697,8 @@ max_d (double lhs, double rhs)
  * to return error information.
  * FRAME is only valid until the next call to alloc_frame ().
  */
-static MONO_NEVER_INLINE void
-interp_exec_method (InterpFrame *frame, ThreadContext *context, FrameClauseArgs *clause_args)
+MONO_NEVER_INLINE void
+mono_interp_exec_method (InterpFrame *frame, ThreadContext *context, FrameClauseArgs *clause_args)
 {
 	InterpMethod *cmethod;
 	ERROR_DECL(error);
@@ -7408,6 +7505,113 @@ MINT_IN_CASE(MINT_BRTRUE_I8_SP) ZEROP_SP(gint64, !=); MINT_IN_BREAK;
 			MINT_IN_BREAK;
 		}
 
+#ifdef HOST_BROWSER
+		MINT_IN_CASE(MINT_TIER_NOP_JITERPRETER) {
+			ip += 3;
+			MINT_IN_BREAK;
+		}
+
+		MINT_IN_CASE(MINT_TIER_PREPARE_JITERPRETER) {
+			if (jiterpreter_traces_enabled) {
+				// We may lose a race with another thread here so we need to use volatile and be careful
+				volatile guint16 *mutable_ip = (volatile guint16*)ip;
+				/*
+				 * prepare_jiterpreter will update the trace's hit count and potentially either JIT it or
+				 *  disable this entry point based on whether it fails to JIT. the hit counting is necessary
+				 *  because a given method may contain many jiterpreter entry points, but some of them will
+				 *  not be actually hit often enough to justify the cost of jitting them. (for example, a
+				 *  trace that only runs inside an unlikely branch for throwing exceptions.)
+				 * thanks to the heuristic that runs during transform.c's codegen, most (95%+) of these
+				 *  entry points will JIT successfully, which will keep the number of NOT_JITTED nops low.
+				 * note: threading doesn't work yet, we will need to broadcast jitted traces to all of our
+				 *  JS workers in order to register them at the appropriate slots in the function pointer
+				 *  table. when growing the function pointer table we will also need to synchronize that.
+				 */
+				JiterpreterThunk prepare_result = mono_interp_tier_prepare_jiterpreter(frame, frame->imethod->method, ip, frame->imethod->jinfo->code_start, frame->imethod->jinfo->code_size);
+				switch ((guint32)(void*)prepare_result) {
+					case JITERPRETER_TRAINING:
+						// jiterpreter still updating hit count before deciding to generate a trace,
+						//  so skip this opcode.
+						ip += 3;
+						break;
+					case JITERPRETER_NOT_JITTED:
+						// Patch opcode to disable it because this trace failed to JIT.
+						mono_memory_barrier();
+						*mutable_ip = MINT_TIER_NOP_JITERPRETER;
+						mono_memory_barrier();
+						ip += 3;
+						break;
+					default:
+						/*
+						 * trace generated. patch opcode to disable it, then write the function
+						 *  pointer, then patch opcode again to turn this trace on.
+						 * we do this to ensure that other threads won't see an ENTER_JITERPRETER
+						 *  opcode that has no function pointer stored inside of it.
+						 * (note that right now threading doesn't work, but it's worth being correct
+						 *  here so that implementing thread support will be easier later.)
+						 */
+						*mutable_ip = MINT_TIER_NOP_JITERPRETER;
+						mono_memory_barrier();
+						*(volatile JiterpreterThunk*)(ip + 1) = prepare_result;
+						mono_memory_barrier();
+						*mutable_ip = MINT_TIER_ENTER_JITERPRETER;
+						ip += 3;
+						break;
+				}
+			} else {
+				ip += 3;
+			}
+
+			MINT_IN_BREAK;
+		}
+
+		MINT_IN_CASE(MINT_TIER_ENTER_JITERPRETER) {
+			JiterpreterThunk thunk = (void*)READ32(ip + 1);
+			gboolean trace_requires_safepoint = FALSE;
+			g_assert(thunk);
+			ptrdiff_t offset = thunk(frame, locals);
+			/*
+			 * The trace signals that we need to perform a safepoint by adding a very
+			 *  large amount to the relative displacement. This is because setting a bit
+			 *  in JS via the | operator doesn't work for negative numbers
+			 */
+			if (offset >= 0xE000000) {
+				offset -= 0xF000000;
+				trace_requires_safepoint = TRUE;
+			}
+			/*
+			 * Verify that the offset returned by the thunk is not total garbage
+			 * FIXME: These constants might actually be too small since a method
+			 *  could have massive amounts of IL - maybe we should disable the jiterpreter
+			 *  for methods that big
+			 */
+			g_assertf((offset >= -0xFFFFF) && (offset <= 0xFFFFF), "thunk returned an obviously invalid offset: %i", offset);
+			if (offset <= 0) {
+				BACK_BRANCH_PROFILE (offset);
+			}
+			if (trace_requires_safepoint) {
+				SAFEPOINT;
+			}
+			ip = (guint16*) (((guint8*)ip) + offset);
+			MINT_IN_BREAK;
+		}
+#else
+		MINT_IN_CASE(MINT_TIER_NOP_JITERPRETER) {
+			g_assert_not_reached ();
+			MINT_IN_BREAK;
+		}
+
+		MINT_IN_CASE(MINT_TIER_PREPARE_JITERPRETER) {
+			g_assert_not_reached ();
+			MINT_IN_BREAK;
+		}
+
+		MINT_IN_CASE(MINT_TIER_ENTER_JITERPRETER) {
+			g_assert_not_reached ();
+			MINT_IN_BREAK;
+		}
+#endif
+
 #if !USE_COMPUTED_GOTO
 		default:
 			interp_error_xsx ("Unimplemented opcode: %04x %s at 0x%x\n", *ip, mono_interp_opname (*ip), GPTRDIFF_TO_INT (ip - frame->imethod->code));
@@ -7605,7 +7809,7 @@ interp_run_finally (StackFrameInfo *frame, int clause_index)
 	// this informs MINT_ENDFINALLY to return to EH
 	*(guint16**)(frame_locals (iframe) + iframe->imethod->clause_data_offsets [clause_index]) = NULL;
 
-	interp_exec_method (iframe, context, &clause_args);
+	mono_interp_exec_method (iframe, context, &clause_args);
 
 	iframe->next_free = next_free;
 	iframe->state.ip = state_ip;
@@ -7656,7 +7860,7 @@ interp_run_filter (StackFrameInfo *frame, MonoException *ex, int clause_index, g
 	clause_args.end_at_ip = (const guint16*)handler_ip_end;
 	clause_args.exec_frame = &child_frame;
 
-	interp_exec_method (&child_frame, context, &clause_args);
+	mono_interp_exec_method (&child_frame, context, &clause_args);
 
 	/* Copy back the updated frame */
 	memcpy (iframe->stack, child_frame.stack, iframe->imethod->locals_size);
@@ -7771,7 +7975,7 @@ interp_run_clause_with_il_state (gpointer il_state_ptr, int clause_index, MonoOb
 	/* Set in mono_handle_exception () */
 	context->has_resume_state = FALSE;
 
-	interp_exec_method (&frame, context, &clause_args);
+	mono_interp_exec_method (&frame, context, &clause_args);
 
 	/* Write back args */
 	sp_args = sp;
@@ -8302,3 +8506,96 @@ mono_ee_interp_init (const char *opts)
 	debugger_enabled = mini_get_debug_options ()->mdb_optimizations;
 #endif
 }
+
+#ifdef HOST_BROWSER
+EMSCRIPTEN_KEEPALIVE int
+mono_jiterp_stackval_to_data (MonoType *type, stackval *val, void *data)
+{
+	return stackval_to_data (type, val, data, FALSE);
+}
+
+EMSCRIPTEN_KEEPALIVE int
+mono_jiterp_stackval_from_data (MonoType *type, stackval *result, const void *data)
+{
+	return stackval_from_data (type, result, data, FALSE);
+}
+
+EMSCRIPTEN_KEEPALIVE int
+mono_jiterp_overflow_check_i4 (gint32 lhs, gint32 rhs, int opcode)
+{
+	switch (opcode) {
+		case MINT_MUL_OVF_I4:
+			if (CHECK_MUL_OVERFLOW (lhs, rhs))
+				return 1;
+		break;
+		case MINT_ADD_OVF_I4:
+			if (CHECK_ADD_OVERFLOW (lhs, rhs))
+				return 1;
+		break;
+	}
+
+	return 0;
+}
+
+EMSCRIPTEN_KEEPALIVE int
+mono_jiterp_overflow_check_u4 (guint32 lhs, guint32 rhs, int opcode)
+{
+	switch (opcode) {
+		case MINT_MUL_OVF_UN_I4:
+			if (CHECK_MUL_OVERFLOW_UN (lhs, rhs))
+				return 1;
+		break;
+		case MINT_ADD_OVF_UN_I4:
+			if (CHECK_ADD_OVERFLOW_UN (lhs, rhs))
+				return 1;
+		break;
+	}
+
+	return 0;
+}
+
+EMSCRIPTEN_KEEPALIVE void
+mono_jiterp_ld_delegate_method_ptr (gpointer *destination, MonoDelegate **source)
+{
+	MonoDelegate *del = *source;
+	if (!del->interp_method) {
+		/* Not created from interpreted code */
+		g_assert (del->method);
+		del->interp_method = mono_interp_get_imethod (del->method);
+	} else if (((InterpMethod*)del->interp_method)->optimized_imethod) {
+		del->interp_method = ((InterpMethod*)del->interp_method)->optimized_imethod;
+	}
+	g_assert (del->interp_method);
+	*destination = imethod_to_ftnptr (del->interp_method, FALSE);
+}
+
+void
+mono_jiterp_check_pending_unwind (ThreadContext *context)
+{
+	return check_pending_unwind (context);
+}
+
+void *
+mono_jiterp_get_context (void)
+{
+	return get_context ();
+}
+
+gpointer
+mono_jiterp_frame_data_allocator_alloc (FrameDataAllocator *stack, InterpFrame *frame, int size)
+{
+	return frame_data_allocator_alloc(stack, frame, size);
+}
+
+gboolean
+mono_jiterp_isinst (MonoObject* object, MonoClass* klass)
+{
+	return mono_interp_isinst (object, klass);
+}
+
+gboolean
+mono_interp_is_method_multicastdelegate_invoke (MonoMethod *method)
+{
+	return is_method_multicastdelegate_invoke (method);
+}
+#endif