--- a/src/mono/sample/wasm/simple-raytracer/Makefile
+++ b/src/mono/sample/wasm/simple-raytracer/Makefile
@@ -1,0 +1,11 @@
+TOP=../../../../..
+
+include ../wasm.mk
+
+ifneq ($(AOT),)
+override MSBUILD_ARGS+=/p:RunAOTCompilation=true
+endif
+
+PROJECT_NAME=Wasm.Browser.SimpleRaytracer.csproj
+
+run: run-browser