… src/mono/sample/wasm/browser-bench/main.js
… src/mono/sample/wasm/simple-raytracer/Program.cs
… src/mono/sample/wasm/simple-raytracer/RayTracer.csproj
… src/mono/sample/wasm/simple-raytracer/index.html
… src/mono/sample/wasm/simple-raytracer/main.js
… src/mono/wasm/runtime/cwraps.ts
… src/mono/wasm/runtime/do-jit-call.wasm
… src/mono/wasm/runtime/do-jit-call.wat
… src/mono/wasm/runtime/dotnet.d.ts
… src/mono/wasm/runtime/driver.c
… src/mono/wasm/runtime/es6/dotnet.es6.lib.js
… src/mono/wasm/runtime/exports-internal.ts
… src/mono/wasm/runtime/exports-linker.ts
… src/mono/wasm/runtime/jiterpreter-interp-entry.ts
… src/mono/wasm/runtime/jiterpreter-jit-call.ts
… src/mono/wasm/runtime/jiterpreter-opcodes.ts
… src/mono/wasm/runtime/jiterpreter-support.ts
… src/mono/wasm/runtime/jiterpreter.ts
… src/mono/wasm/runtime/net6-legacy/export-types.ts
… src/mono/wasm/runtime/run.ts
… src/mono/wasm/runtime/types/emscripten.ts
… src/mono/wasm/wasm.proj