@kg 2022-10-07
PR description updated to cover the new 'jiterpreter trampolines', which are jitted replacements for the AOT interp_entry wrappers. The now-sabotaged raytracer sample in the PR forces aot/interp transitions during execution due to runtime generics and delegates, so it shows wildly different timings in each of the 4 possible execution modes (regular AOT, AOT+trampolines, regular interpreter, interpreter+jiterpreter+trampolines). There are opportunities to JIT other wrappers to further improve the performance of this test in AOT mode, I think - it spends a lot of time on overhead in various places instead of actually doing work, and before I sabotaged it it was MUCH faster in AOT.

Comparison timings follow (best time selected from a set of runs in each case). Note that if you are running it for measurement, *do not* open devtools until all your runs are done, because devtools deoptimize JS and wasm code. If you want to profile it, start the profiler and then *reload the page* while the profiler is running.

```
/// aot, no optimizations
Rendering finished in 7191 ms
// jiterpreter produced 0 traces from 0 candidates, and 0 trampolines
// time spent: 0ms generating, 0ms compiling wasm


/// aot with traces and trampolines enabled
Rendering finished in 6863 ms
// jiterpreter produced 0 traces from 2 candidates, and 2 trampolines
// time spent: 3ms generating, 2ms compiling wasm


/// interpreter, no optimizations
Rendering finished in 4868 ms
// jiterpreter produced 0 traces from 0 candidates, and 0 trampolines
// time spent: 0ms generating, 0ms compiling wasm


/// interpreter with traces and trampolines enabled
Rendering finished in 4070 ms
// jiterpreter produced 3 traces from 5 candidates, and 0 trampolines
// time spent: 6ms generating, 0ms compiling wasm
```