@kg 2022-10-13
I reworked the size threshold logic used to select where to insert traces, and implemented a C-side heuristic that is able to make fairly accurate guesses about where traces should go which reduces the overhead introduced by the jiterpreter when it fails to JIT a method. The timings for the toy raytracer testcase are still roughly as above, but the addition of the heuristic means less time is spent generating code in the runtime test suite, so the jiterpreter is now a measurable perf improvement for it. Some summary timings for fresh runs of the toy raytracer, pavel's [more complex raytracer sample](https://github.com/pavelsavara/dotnet-wasm-raytracer) (w/o SIMD), and the system.runtime test suite (which currently is incompatible with AOT):

|name|interpreter (msec)|JITerpreter (msec)|% of baseline|AOT (msec)|AOT+JIT (msec)|% of baseline|
|-|-|-|-|-|-|-|
|`System.Runtime.Tests`|140847|125983|**89.44%**|N/A|N/A|N/A|
|Sabotaged raytracer|9321|6681|**71.67%**|12345|11645|**94.32%**|
|Pavel's raytracer|20462|13534|**66.14%**|2267|N/A|N/A|