@pavelsavara 2022-10-18
I think we want to make it Mono component.
Also make it optional behind rollup flag in the typeScript (if we don't rewrite it in C).
See `MonoWasmThreads` in the .ts code.