@kg 2022-10-29
PR updated with an experimental JIT for do_jit_call that replaces a bunch of the wrapper layers with a single JS function + jitted wasm function pair that passes data directly from the interpreter stack to the target jitcode (or wrapper, depending). Current performance improvement is modest but definitely there for AOT browser-bench and the sabotaged raytracer, since both spend a sizable amount of time moving between interp and AOT code. (This already has the optimized interp_entry JIT enabled along with the jiterpreter.)

| measurement | time (baseline) | time (optimized) | relative time (%) |
|-:|-:|-:|-:|
|                  Json, small serialize |     0.0283ms |     0.0277ms | **97.8%** |
|                Json, small deserialize |     0.0497ms |     0.0488ms | **98.1%** |
|                  Json, large serialize |     7.6641ms |     7.4977ms | **97.8%** |
|                Json, large deserialize |    14.1278ms |    13.7778ms | **97.5%** |
|               Sabotaged raytracer, AOT |     11943 ms |     10824 ms | **90.6%** |

I believe it's possible to make them much faster.