@kg 2022-11-02
The NDPin test failure in https://helixre107v0xdeko0k025g8.blob.core.windows.net/dotnet-runtime-refs-pull-76477-merge-2867209773cb4d30b9/PayloadGroup0/1/console.c9cfd27c.log?helixlogtype=result is real, I'm looking into it (though I don't know how to run the test harness it's from since it's generated, the test code fails when pasted into a browser sample.)