@pavelsavara 2022-11-09
I would love this to be optional part of `dotnet.js` behind a feature flag. It may be in next PR tho.