@kg 2022-11-09
> I would love this to be optional part of `dotnet.js` behind a feature flag. It may be in next PR tho.

With the options interface in this PR, we can set the default to off and then you use a runtime flag to turn it on. See the ```.withRuntimeOptions(["--jiterpreter-enable-stats"])``` in browser-bench, there are flags to enable/disable parts of the jiterpreter as well.