@pavelsavara 2022-11-09
> With the options interface in this PR, we can set the default to off and then you use a runtime flag to turn it on. 

I mean, to link it out via rollup. 
Or even better, this would be great candidate to make it separate ES6 module and dynamically `import()` it only when the config flag is set.