@kg 2022-11-10
Here are some updated timings for various scenarios, since some of the performance characteristics have changed since the previous timings (a few key reasons are that batching the JIT operation can cause slower code to run for a period of time - this slows down console tests but not most real applications, and that I had to revert the inlining threshold).

System.Runtime.Tests now passes with AOT enabled, but while the jiterpreter is active for it, the fact that the suite runs synchronously means that the improvement is small enough to be within the margin of error, so I omitted it.

| name | baseline (ms) | jiterpreter (ms) | relative time (%) |
|-:|-:|-:|-:|
|                System.Runtime.Tests (interp) |   138651.7 |   129579.6 | **93.46%** |
|                  Json, large serialize (AOT) |     8.1799 |     7.6687 | **93.75%** |
|                Json, large deserialize (AOT) |    14.3750 |    14.1799 | **98.64%** |
|             JSInterop, LegacyExportInt (AOT) |     8.4975 |     7.9102 | **93.09%** |
|                 JSInterop, JSExportInt (AOT) |     6.4096 |     5.3799 | **83.94%** |
|                 JSInterop, JSImportInt (AOT) |     0.7562 |     0.6748 | **89.24%** |
|                   mixed-mode raytracer (AOT) |      12751 |       9861 | **77.34%** |
|               Json, small serialize (interp) |     0.2517 |     0.2093 | **83.15%** |
|             Json, small deserialize (interp) |     0.3567 |     0.2901 | **81.33%** |
|               Json, large serialize (interp) |    72.5972 |    57.2889 | **78.91%** |
|             Json, large deserialize (interp) |    97.6667 |    79.5303 | **81.43%** |
|          JSInterop, LegacyExportInt (interp) |     4.8936 |     4.8633 | **99.38%** |
|              JSInterop, JSExportInt (interp) |     3.6925 |     3.2976 | **89.31%** |
|                mixed-mode raytracer (interp) |       9271 |       7024 | **75.76%** |
|                pavel's raytracer    (interp) |      21353 |      13810 | **64.67%** |